Update and rename Cant.java to CarNerdMainPage.java
